feedback: 지하철역, 노선 관련 인수테스트 추상 클래스 분리
