+++ src/test/java/nextstep/subway/station/BaseLineAcceptanceTest.java
@@ -4,38 +4,12 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.StationRequest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 
 @Sql("/truncate.sql")
-public abstract class BaseAcceptanceTest {
-    private static final String STATIONS_URI = "/stations";
+public abstract class BaseLineAcceptanceTest {
     private static final String LINES_URI = "/lines";
-
-    protected ExtractableResponse<Response> createStationRequest(String stationName) {
-        return RestAssured.given().log().all()
-                .body(StationRequest.from(stationName))
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post(STATIONS_URI)
-                .then().log().all()
-                .extract();
-    }
-
-    protected ExtractableResponse<Response> findStationsRequest() {
-        return RestAssured.given().log().all()
-                .when().get(STATIONS_URI)
-                .then().log().all()
-                .extract();
-    }
-
-    protected ExtractableResponse<Response> deleteStationRequest(int stationId) {
-        return RestAssured.given().log().all()
-                .pathParam("id", stationId)
-                .when().delete(STATIONS_URI + "/{id}")
-                .then().log().all()
-                .extract();
-    }
 
     protected ExtractableResponse<Response> createLineRequest(LineRequest lineRequest) {
         return RestAssured.given().log().all()
